Extract settings sync panel
